+++ test/test_extensions.rb
@@ -64,27 +64,4 @@ class TestExtensions < MiniTest::Unit::TestCase
       SomeClass.delay.doit(Date.today)
     end
   end
-
-  describe 'sidekiq rails extensions configuration' do
-    before do
-      @options = Sidekiq.options
-    end
-
-    after do
-      Sidekiq.options = @options
-    end
-
-    it 'should set enable_rails_extensions option to true by default' do
-      assert Sidekiq.options[:enable_rails_extensions]
-    end
-
-    it 'should extend ActiveRecord and ActiveMailer if enable_rails_extensions is true' do
-      assert Sidekiq.hook_rails!
-    end
-
-    it 'should not extend ActiveRecord and ActiveMailer if enable_rails_extensions is false' do
-      Sidekiq.options = { :enable_rails_extensions => false }
-      refute Sidekiq.hook_rails!
-    end
-  end
 end